add initial deduplication
